requirements.txt
flask
PyPDF2
python-docx
scikit-learn
spacy
